[fix] avoid DateTime addition int overflows
spec-ed DateTime#- part still failing, for now

adjusted from GH-5369

+++ core/src/main/java/org/jruby/ext/date/RubyDate.java
@@ -1169,8 +1169,8 @@ public class RubyDate extends RubyObject {
         RubyNumeric sub = (RubyNumeric) res.eltInternal(1);
         //if ( sub.isZero() ) sub = RubyFixnum.zero(runtime); // avoid Rational(0, 1)
         RubyNumeric sub_millis = (RubyNumeric) subMillis(context.runtime).op_plus(context, sub);
-        int subNum = sub_millis.numerator(context).convertToInteger().getIntValue();
-        int subDen = sub_millis.denominator(context).convertToInteger().getIntValue();
+        long subNum = sub_millis.numerator(context).convertToInteger().getLongValue();
+        long subDen = sub_millis.denominator(context).convertToInteger().getLongValue();
         if (subNum / subDen >= 1) { // sub_millis >= 1
             subNum -= subDen; ms += 1; // sub_millis -= 1
         }
